requirements.txt
streamlit
pandas
scikit-learn
opencv-python
pillow
numpy